# docs/swagger/swagger.json
{
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Qredo Signing Agent service interacts with [Partner API](https://developers.qredo.com/partner-api/api/swagger/) to register a Signing Agent to automate approvals according to your custody policy. \u003cbr/\u003e\nAuthentication and encryption are required; set up your [API key and secret in the Qredo Web App](https://developers.qredo.com/signing-agent/v2-signing-agent/get-started/). \u003cbr/\u003e",
    "title": "Qredo Signing Agent V2",
    "contact": {
      "name": "Qredo API Services",
      "url": "https://www.qredo.com",
      "email": "support@qredo.com"
    },
    "license": {
      "name": "APACHE 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "0.1.0"
  },
  "host": "localhost:8007",
  "basePath": "/api/v1",
  "paths": {
    "/client": {
      "get": {
        "tags": [
          "client"
        ],
        "summary": "Returns AgentID if it's configured",
        "operationId": "ClientsList",
        "responses": {
          "200": {
            "description": "ClientListResponse",
            "schema": {
              "$ref": "#/definitions/ClientListResponse"
            }
          }
        }
      }
    },
    "/client/action/{action_id}": {
      "put": {
        "tags": [
          "action"
        ],
        "summary": "Approve a transaction",
        "operationId": "actionApprove",
        "parameters": [
          {
            "type": "string",
            "description": "the id of the transaction that is received from the feed",
            "name": "action_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "GenericResponse",
            "schema": {
              "$ref": "#/definitions/GenericResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "action"
        ],
        "summary": "Reject a transaction",
        "operationId": "actionReject",
        "parameters": [
          {
            "type": "string",
            "description": "the id of the transaction that is received from the feed",
            "name": "action_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "GenericResponse",
            "schema": {
              "$ref": "#/definitions/GenericResponse"
            }
          }
        }
      }
    },
    "/client/feed": {
      "post": {
        "produces": [
          "application/json"
        ],
        "schemes": [
          "ws",
          "wss"
        ],
        "tags": [
          "client"
        ],
        "summary": "Get approval requests Feed (via websocket) from Qredo Backend",
        "operationId": "ClientFeed",
        "responses": {
          "200": {
            "description": "ClientFeedResponse",
            "schema": {
              "$ref": "#/definitions/ClientFeedResponse"
            }
          }
        }
      }
    },
    "/client/register": {
      "post": {
        "tags": [
          "client"
        ],
        "summary": "Register a new agent",
        "operationId": "RegisterAgent",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/ClientRegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ClientFullRegisterResponse",
            "schema": {
              "$ref": "#/definitions/ClientFullRegisterResponse"
            }
          }
        }
      }
    },
    "/healthcheck/config": {
      "get": {
        "tags": [
          "healthcheck"
        ],
        "summary": "Check application configuration.",
        "operationId": "HealthcheckConfig",
        "responses": {
          "200": {
            "description": "ConfigResponse",
            "schema": {
              "$ref": "#/definitions/ConfigResponse"
            }
          }
        }
      }
    },
    "/healthcheck/status": {
      "get": {
        "tags": [
          "healthcheck"
        ],
        "summary": "Check application status.",
        "operationId": "HealthcheckStatus",
        "responses": {
          "200": {
            "description": "HealthCheckStatusResponse",
            "schema": {
              "$ref": "#/definitions/HealthCheckStatusResponse"
            }
          }
        }
      }
    },
    "/healthcheck/version": {
      "get": {
        "tags": [
          "healthcheck"
        ],
        "summary": "Check application version.",
        "operationId": "HealthcheckVersion",
        "responses": {
          "200": {
            "description": "VersionResponse",
            "schema": {
              "$ref": "#/definitions/VersionResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AWSConfig": {
      "type": "object",
      "title": "AWSConfig based signing-agent config. Used when Base StoreType is aws.",
      "properties": {
        "Region": {
          "type": "string"
        },
        "SecretName": {
          "type": "string"
        }
      }
    },
    "AutoApprove": {
      "type": "object",
      "properties": {
        "Enabled": {
          "type": "boolean"
        },
        "RetryInterval": {
          "type": "integer",
          "format": "int64"
        },
        "RetryIntervalMax": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "Base": {
      "type": "object",
      "properties": {
        "PIN": {
          "type": "integer",
          "format": "int64"
        },
        "QredoAPI": {
          "type": "string"
        }
      }
    },
    "ClientFeedResponse": {
      "type": "object",
      "properties": {
        "coreClientID": {
          "description": "The ID of the agent",
          "type": "string",
          "example": "98cTMMSPrDdcDDVU8idhuJGK2U1P4vmQcsp8wnED8pPR"
        },
        "expireTime": {
          "description": "The time that the transaction will expire, utc unix time",
          "type": "integer",
          "format": "int64",
          "example": 1676184187
        },
        "id": {
          "description": "The ID of the transaction",
          "type": "string",
          "example": "2IXwq4klvWbnPf1YaAc1XD85jJX"
        },
        "status": {
          "description": "The status of the transaction",
          "type": "string",
          "example": "pending"
        },
        "timestamp": {
          "description": "The time that the transaction was started, utc unix time",
          "type": "integer",
          "format": "int64",
          "example": 1670341423
        },
        "type": {
          "description": "The type of the transaction",
          "type": "string",
          "example": "ApproveWithdraw"
        }
      }
    },
    "ClientFullRegisterResponse": {
      "type": "object",
      "properties": {
        "agentId": {
          "type": "string"
        },
        "feedUrl": {
          "type": "string"
        }
      }
    },
    "ClientListResponse": {
      "type": "array",
      "items": {
        "type": "string"
      }
    },
    "ClientRegisterRequest": {
      "type": "object",
      "properties": {
        "apikey": {
          "description": "The api key for the partner api",
          "type": "string",
          "example": "eyJrZXkiOiJHM0Fo... (truncated)"
        },
        "base64privatekey": {
          "description": "The base64 encoded private key pem of which the public key has been registered in the partner api",
          "type": "string",
          "example": "LS0tLS1CRUdJTiBS... (truncated)"
        },
        "name": {
          "description": "The name of the agent",
          "type": "string",
          "example": "test-agent"
        }
      }
    },
    "ConfigResponse": {
      "type": "object",
      "properties": {
        "AutoApprove": {
          "$ref": "#/definitions/AutoApprove"
        },
        "Base": {
          "$ref": "#/definitions/Base"
        },
        "HTTP": {
          "$ref": "#/definitions/HttpSettings"
        },
        "LoadBalancing": {
          "$ref": "#/definitions/LoadBalancing"
        },
        "Logging": {
          "$ref": "#/definitions/Logging"
        },
        "Store": {
          "$ref": "#/definitions/Store"
        },
        "Websocket": {
          "$ref": "#/definitions/WebSocketConf"
        }
      }
    },
    "GenericResponse": {
      "type": "object",
      "properties": {
        "Code": {
          "description": "The result code of the request",
          "type": "integer",
          "format": "int64",
          "example": 200
        },
        "Msg": {
          "description": "The result message of the request",
          "type": "string",
          "example": "OK"
        }
      }
    },
    "HealthCheckStatusResponse": {
      "type": "object",
      "properties": {
        "WebSocket": {
          "$ref": "#/definitions/WebsocketStatus"
        }
      }
    },
    "HttpSettings": {
      "type": "object",
      "properties": {
        "Addr": {
          "type": "string"
        },
        "CORSAllowOrigins": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "LogAllRequests": {
          "type": "boolean"
        },
        "TLS": {
          "$ref": "#/definitions/TLSConfig"
        }
      }
    },
    "LoadBalancing": {
      "type": "object",
      "properties": {
        "ActionIDExpirationSec": {
          "type": "integer",
          "format": "int64"
        },
        "Enable": {
          "type": "boolean"
        },
        "OnLockErrorTimeOutMs": {
          "type": "integer",
          "format": "int64"
        },
        "RedisConfig": {
          "$ref": "#/definitions/RedisConfig"
        }
      }
    },
    "Logging": {
      "type": "object",
      "properties": {
        "Format": {
          "type": "string"
        },
        "Level": {
          "type": "string"
        }
      }
    },
    "OciConfig": {
      "type": "object",
      "properties": {
        "Compartment": {
          "type": "string"
        },
        "ConfigSecret": {
          "type": "string"
        },
        "SecretEncryptionKey": {
          "type": "string"
        },
        "Vault": {
          "type": "string"
        }
      }
    },
    "RedisConfig": {
      "type": "object",
      "properties": {
        "DB": {
          "type": "integer",
          "format": "int64"
        },
        "Host": {
          "type": "string"
        },
        "Password": {
          "type": "string"
        },
        "Port": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "Store": {
      "type": "object",
      "properties": {
        "AwsConfig": {
          "$ref": "#/definitions/AWSConfig"
        },
        "FileConfig": {
          "type": "string"
        },
        "OciConfig": {
          "$ref": "#/definitions/OciConfig"
        },
        "Type": {
          "type": "string"
        }
      }
    },
    "TLSConfig": {
      "type": "object",
      "properties": {
        "CertFile": {
          "type": "string"
        },
        "Enabled": {
          "type": "boolean"
        },
        "KeyFile": {
          "type": "string"
        }
      }
    },
    "VersionResponse": {
      "type": "object",
      "properties": {
        "BuildDate": {
          "type": "string"
        },
        "BuildType": {
          "type": "string"
        },
        "BuildVersion": {
          "type": "string"
        }
      }
    },
    "WebSocketConf": {
      "type": "object",
      "properties": {
        "PingPeriod": {
          "type": "integer",
          "format": "int64"
        },
        "PongWait": {
          "type": "integer",
          "format": "int64"
        },
        "QredoWebsocket": {
          "type": "string"
        },
        "ReadBufferSize": {
          "type": "integer",
          "format": "int64"
        },
        "ReconnectInterval": {
          "type": "integer",
          "format": "int64"
        },
        "ReconnectTimeOut": {
          "type": "integer",
          "format": "int64"
        },
        "WriteBufferSize": {
          "type": "integer",
          "format": "int64"
        },
        "WriteWait": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "WebsocketStatus": {
      "type": "object",
      "properties": {
        "ConnectedClients": {
          "type": "integer",
          "format": "uint32"
        },
        "LocalFeedUrl": {
          "type": "string"
        },
        "ReadyState": {
          "type": "string"
        },
        "RemoteFeedUrl": {
          "type": "string"
        }
      }
    }
  }
}